Binary code generation.
Powering through the translator so I can start working on the virtual machine.
This does the absolute least it can do to make interesting binaries: it can't
handle local variables (static only, thanks) and it can't handle function calls,
but all operations are now translated into a binary format. Constants can't be
used for anything that doesn't fit into a byte -- they'll just be truncated. But
now I get to start working on the VM, so whatever.