Support '#include' preprocessor directive.
Note that because of how functions are called at the moment, including headers
doesn't do a ton of good unless they contain the function definitions, not just
stubs. Also, you can't pass more than one source file to the compiler at the
moment, so you can't use other C code. Given that, I am adding "cvmlib", a dinky
implementation of the bits of libc that I use. For now, that's just putc,
because I'm tired of editing the assembly to add PRINT instructions.